tests for authentication and file endpoints
