+++ gateway/sds_gateway/api_methods/tests/test_file_endpoints.py
@@ -1,5 +1,3 @@
-# Create your tests here.
-# tests.py
 from pathlib import Path
 
 from django.contrib.auth import get_user_model
@@ -10,6 +8,7 @@ from rest_framework.test import APIClient
 from rest_framework.test import APITestCase
 
 from sds_gateway.api_methods.serializers.file_serializers import FilePostSerializer
+from sds_gateway.users.models import UserAPIKey
 
 User = get_user_model()
 
@@ -21,6 +20,12 @@ class FileTestCases(APITestCase):
             email="testuser@example.com",
             password="testpassword",  # noqa: S106
         )
+
+        self.api_key, self.key = UserAPIKey.objects.create_key(
+            name="test-key",
+            user=self.user,
+        )
+
         # Create a test file object
         self.file_obj = SimpleUploadedFile(
             "testfile.txt",
@@ -29,7 +34,7 @@ class FileTestCases(APITestCase):
         )
         file_data = {
             "file": self.file_obj,
-            "directory": "test_directory",
+            "directory": "/absolute/path/to/files",
             "media_type": "text/plain",
         }
         serializer = FilePostSerializer(
@@ -39,9 +44,16 @@ class FileTestCases(APITestCase):
         serializer.is_valid(raise_exception=True)
         self.file = serializer.save()
 
-        self.client.force_authenticate(user=self.user)
+        self.client.credentials(HTTP_AUTHORIZATION=f"Api-Key {self.key}")
         self.list_url = reverse("api:files-list")
         self.detail_url = reverse("api:files-detail", args=[self.file.uuid])
+        self.contents_check_url = reverse("api:check_contents_exist")
+
+    def tearDown(self):
+        # Delete the test file if it exists
+        test_file_path = Path("testfile.txt")
+        if test_file_path.exists():
+            test_file_path.unlink()
 
     def test_create_file(self):
         with Path("testfile.txt").open("w") as file:
@@ -50,7 +62,7 @@ class FileTestCases(APITestCase):
         with Path("testfile.txt").open("rb") as file:
             data = {
                 "file": file,
-                "directory": "test_directory",
+                "directory": "/absolute/path/to/files",
                 "media_type": "text/plain",
             }
             response = self.client.post(self.list_url, data, format="multipart")
@@ -58,7 +70,7 @@ class FileTestCases(APITestCase):
         assert response.status_code == status.HTTP_201_CREATED
 
     def test_retrieve_file(self):
-        response = self.client.get(reverse("api:files-detail", args=[self.file.uuid]))
+        response = self.client.get(self.detail_url)
         assert response.status_code == status.HTTP_200_OK
 
     def test_retrieve_latest_file(self):
@@ -88,3 +100,16 @@ class FileTestCases(APITestCase):
             self.detail_url,
         )
         assert response.status_code == status.HTTP_204_NO_CONTENT
+
+    # check where file can be updated
+    def test_file_contents_check(self):
+        data = {
+            "file": self.file_obj,
+            "directory": self.file.directory,
+            "name": self.file.name,
+        }
+        response = self.client.post(self.contents_check_url, data, format="multipart")
+        assert response.status_code == status.HTTP_200_OK
+        assert "file_contents_exist_for_user" in response.data
+        assert "file_exists_in_tree" in response.data
+        assert "user_mutable_attributes_differ" in response.data
